+++ packages/guide/src/app/journal/journal-routing.module.ts
@@ -4,6 +4,7 @@ import { JournalComponent } from './journal.component';
 import { CaughtComponent } from './components/caught/caught.component';
 import { FoundComponent } from './components/found/found.component';
 import { ProduceComponent } from './components/produce/produce.component';
+import { AchievementsComponent } from "./components/achievements/achievements.component";
 
 const routes: Routes = [
     {
@@ -22,6 +23,8 @@ const routes: Routes = [
             {path: 'found/:tabName', component: FoundComponent, title: 'Found - Journal'},
             {path: 'produce', redirectTo: 'produce/', pathMatch: 'full'},
             {path: 'produce/:tabName', component: ProduceComponent, title: 'Produce - Journal'},
+            {path: 'achievements', redirectTo: 'achievements/', pathMatch: 'full'},
+            {path: 'achievements/:achievementId', component: AchievementsComponent, title: 'Achievements - Journal'},
         ]
     },
 ];